+++ frontend/src/images/chatbot-logo.png
